# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: OHB Reinraum Sensor Dashboard — API
  version: 1.0.0
  description: |-
    Schnittstelle des Reinraum-Überwachungssystems.

    **Fehlerformat:** Alle Fehler antworten mit `{ error: { code, message, details? }, requestId }`.
    Die `requestId` steht zusätzlich im Antwort-Header `X-Request-Id` und erlaubt die
    Zuordnung zu den Logzeilen des Servers.

    **Zeitangaben:** durchgängig ISO 8601 mit Zeitzonenangabe (`timestamptz`).

    **Zeitraum-Versionierung:** Zuordnungen und Schwellenwerte werden nie überschrieben.
    Eine Änderung schließt den bisherigen Gültigkeitszeitraum und öffnet einen neuen.
    Auswertungen beziehen sich deshalb immer auf den Stand zum jeweiligen Zeitpunkt.
servers:
  - url: /
    description: Gleiche Herkunft (hinter nginx)
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Stabiler Fehlercode für Clients
            message:
              type: string
            details: {}
          required:
            - code
            - message
        requestId:
          type: string
          description: Korrelations-ID, auch im Header X-Request-Id
      required:
        - error
        - requestId
...
    Cleanroom:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
        - id
        - name
    CreateCleanroomBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
      required:
        - name
...
    RenameSensorBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
      required:
        - name
    Assignment:
      type: object
      properties:
        id:
          type: integer
        sensor_uuid:
          type: string
        sensor_name:
          type:
            - string
            - "null"
        gateway_id:
          type:
            - string
            - "null"
        cleanroom_id:
          type: integer
        cleanroom_name:
          type: string
        valid_during:
          description: Gültigkeitszeitraum als tstzrange
      required:
        - id
        - sensor_uuid
        - sensor_name
        - gateway_id
        - cleanroom_id
        - cleanroom_name
    AssignmentHistoryEntry:
      type: object
      properties:
        id:
          type: integer
        sensor_uuid:
          type: string
        cleanroom_id:
          type: integer
        cleanroom_name:
          type: string
        valid_from:
          anyOf:
            - type: string
              format: date-time
            - type: string
            - type: "null"
        valid_to:
          anyOf:
            - type: string
              format: date-time
            - type: string
            - type: "null"
          description: null = bis auf Weiteres
      required:
        - id
        - sensor_uuid
        - cleanroom_id
        - cleanroom_name
        - valid_from
        - valid_to
    CreateAssignmentBody:
      type: object
      properties:
        sensor_uuid:
          type: string
          format: uuid
          description: Sensor-Kennung im Format 8-4-4-4-12
        cleanroom_id:
          type: integer
          exclusiveMinimum: 0
      required:
        - sensor_uuid
        - cleanroom_id
...
    DataPoint:
      type: object
      properties:
        time:
          anyOf:
            - type: string
              format: date-time
            - type: string
        quantity:
          type: string
        unit:
          type: string
        value:
          type: number
      required:
        - time
        - quantity
        - unit
        - value
...
paths:
...
  /api/cleanrooms:
    get:
      tags:
        - Reinräume
      summary: Alle Reinräume auflisten
      responses:
        "200":
          description: Liste der Reinräume
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cleanroom"
    post:
      tags:
        - Reinräume
      summary: Reinraum anlegen
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCleanroomBody"
      responses:
        "201":
          description: Angelegter Reinraum
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cleanroom"
        "400":
          description: Ungültige Eingabe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Konflikt mit dem aktuellen Zustand
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/cleanrooms/{id}:
    delete:
      tags:
        - Reinräume
      summary: Reinraum löschen
      description: Laufende Zuordnungen werden abgeschlossen, Verletzungen behalten ihre Historie und
        verlieren lediglich den Raumbezug.
      parameters:
        - schema:
            type: integer
            exclusiveMinimum: 0
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Gelöscht
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                required:
                  - deleted
        "400":
          description: Ungültige Eingabe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Nicht gefunden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/sensors/{sensor_uuid}:
    patch:
      tags:
        - Sensoren
      summary: Sensor umbenennen
      description: Ein hier vergebener Name wird vom Ingest nie durch einen vom Gerät gemeldeten Namen
        überschrieben.
      parameters:
        - schema:
            type: string
            format: uuid
            description: Sensor-Kennung im Format 8-4-4-4-12
          required: true
          description: Sensor-Kennung im Format 8-4-4-4-12
          name: sensor_uuid
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameSensorBody"
      responses:
        "200":
          description: Aktualisierter Sensor
          content:
            application/json:
              schema:
                type: object
                properties:
                  sensor_uuid:
                    type: string
                  name:
                    type: string
                  gateway_id:
                    type:
                      - string
                      - "null"
                  event_driven:
                    type: integer
                    description: 0 = zyklisch, 1 = event-getrieben
                  created_at:
                    anyOf:
                      - type: string
                        format: date-time
                      - type: string
                  cleanroom_id:
                    type:
                      - integer
                      - "null"
                  cleanroom_name:
                    type:
                      - string
                      - "null"
                  quantities:
                    type: array
                    items:
                      type: string
                    description: Bisher gemessene Größen
        "400":
          description: Ungültige Eingabe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Nicht gefunden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Sensoren
      summary: Sensor entfernen
      description: Entfernt Register-Eintrag, Zuordnungen, Schwellenwerte und Verletzungen. Die Messdaten
        selbst bleiben erhalten — sie gehören zur Historie des Reinraums.
      parameters:
        - schema:
            type: string
            format: uuid
            description: Sensor-Kennung im Format 8-4-4-4-12
          required: true
          description: Sensor-Kennung im Format 8-4-4-4-12
          name: sensor_uuid
          in: path
      responses:
        "200":
          description: Entfernt
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                required:
                  - deleted
        "400":
          description: Ungültige Eingabe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Nicht gefunden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/assignments:
    get:
      tags:
        - Zuordnungen
      summary: Aktuell gültige Zuordnungen
      responses:
        "200":
          description: Zuordnungen
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      tags:
        - Zuordnungen
      summary: Sensor einem Reinraum zuordnen
      description: Schließt die bisherige Zuordnung ab und öffnet eine neue — beides in einer Transaktion,
        sodass weder Lücke noch Überschneidung entstehen kann.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentBody"
      responses:
        "201":
          description: Neue Zuordnung
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  sensor_uuid:
                    type: string
                  sensor_name:
                    type:
                      - string
                      - "null"
                  gateway_id:
                    type:
                      - string
                      - "null"
                  cleanroom_id:
                    type: integer
                  cleanroom_name:
                    type: string
                  valid_during:
                    description: Gültigkeitszeitraum als tstzrange
        "400":
          description: Ungültige Eingabe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Konflikt mit dem aktuellen Zustand
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/assignments/history:
    get:
      tags:
        - Zuordnungen
      summary: Vollständige Raum-Historie eines Sensors
      description: "`valid_to = null` bedeutet: bis auf Weiteres gültig."
      parameters:
        - schema:
            type: string
            format: uuid
            description: Sensor-Kennung im Format 8-4-4-4-12
          required: true
          description: Sensor-Kennung im Format 8-4-4-4-12
          name: sensor_uuid
          in: query
      responses:
        "200":
          description: Historie, neueste zuerst
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssignmentHistoryEntry"
        "400":
          description: Ungültige Eingabe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/sensordata:
    get:
      tags:
        - Messdaten
      summary: Zeitreihe eines Kanals
      description: Ohne Zeitangaben werden die letzten 24 Stunden geliefert. Bei Erreichen von `limit`
        werden die JÜNGSTEN Werte zurückgegeben; die Ausgabe ist aufsteigend sortiert.
      parameters:
        - schema:
            type: string
            format: uuid
            description: Sensor-Kennung im Format 8-4-4-4-12
          required: true
          description: Sensor-Kennung im Format 8-4-4-4-12
          name: sensor_uuid
          in: query
        - schema:
            type: string
            minLength: 1
            maxLength: 64
            description: Messgröße
          required: true
          description: Messgröße
          name: quantity
          in: query
        - schema:
            type: string
            format: date-time
            description: "Standard: vor 24 Stunden"
          required: false
          description: "Standard: vor 24 Stunden"
          name: from
          in: query
        - schema:
            type: string
            format: date-time
            description: "Standard: jetzt"
          required: false
          description: "Standard: jetzt"
          name: to
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 5000
            default: 500
          required: false
          name: limit
          in: query
      responses:
        "200":
          description: Messwerte, älteste zuerst
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataPoint"
        "400":
          description: Ungültige Eingabe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...